fix(speak): bad link (#382)
If you are an editor of the form, it opens the editor. It also has a
"Request to Edit" button appear. This is the shareable link version.

<!-- 👋 Hi, thanks for sending a PR to boston-ts-website! 💖.
Please fill out all fields below and make sure each item is true and [x]
checked.
Otherwise we may not be able to review your PR. -->

## PR Checklist

- [x] Steps in
[CONTRIBUTING.md](https://github.com/JoshuaKGoldberg/boston-ts-website/blob/main/.github/CONTRIBUTING.md)
were taken

## Overview

See commit message

<!-- Description of what is changed and how the code change does that.
-->

🐢